+++ settings.gradle.kts
@@ -1,3 +1,2 @@
 
 rootProject.name = "dynamokt"
-
