+++ src/main/kotlin/com/aseemsavio/dynamokt/annotations/codegen/poet/Utils.kt
@@ -113,7 +113,4 @@ internal fun FunSpec.Builder.addStatementAccordingToType(prop: KotlinProperty) =
             )
         )
     }
-
 }
-
-
