+++ src/main/kotlin/com/aseemsavio/dynamokt/annotations/visitors/DynamoKtAnnotationVisitor.kt
@@ -1,7 +1,7 @@
 package com.aseemsavio.dynamokt.annotations.visitors
 
 import com.aseemsavio.dynamokt.annotations.codegen.dataClassInfo
-import com.aseemsavio.dynamokt.annotations.codegen.poet.generateAttributeMapCode
+import com.aseemsavio.dynamokt.annotations.codegen.poet.generateDynamoKtCode
 import com.google.devtools.ksp.processing.CodeGenerator
 import com.google.devtools.ksp.processing.KSPLogger
 import com.google.devtools.ksp.symbol.KSClassDeclaration
@@ -15,7 +15,7 @@ class DynamoKtAnnotationVisitor(
     override fun visitClassDeclaration(classDeclaration: KSClassDeclaration, data: Unit) {
         val dataClassInfo = classDeclaration.dataClassInfo
 
-        generateAttributeMapCode(
+        generateDynamoKtCode(
             dataClassInfo = dataClassInfo,
             classDeclaration = classDeclaration,
             codeGenerator = codeGenerator,
